requirements.txt
streamlit==1.28.1
google-analytics-data==0.18.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
plotly==5.17.0
python-dateutil==2.8.2
